refactor: reorganize file structures in XCode project
