+++ Workout Spinner WatchKit Extension/Views/General Components/SpinnerRotationModifier.swift
