+++ Workout Spinner WatchKit Extension/Views/General Components/Text ViewModifiers.swift
@@ -1,0 +1,18 @@
+//
+//  Text ViewModifiers.swift
+//  Workout Spinner WatchKit Extension
+//
+//  Created by Joshua on 12/3/20.
+//  Copyright © 2020 Joshua Cook. All rights reserved.
+//
+
+import SwfitUI
+
+struct DoneButtonText: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .font(.system(.body, design: .rounded))
+            .foregroundColor(.white)
+            .frame(minWidth: 0, maxWidth: .infinity)
+    }
+}
